feat(mcp): move the landing page from /teachers/mcp to /mcp
The MCP server is moving to mcp.thenational.academy, which frees the /mcp path
on www for this page. Git records the whole move as renames, so the diff is the
import paths and the pa11y/Percy test URL.

Still inside the (core) route group, so it keeps TopNav, main and the site
footer. No redirect from /teachers/mcp: the page has never shipped, so nothing
links to the old path.

⚠️ Merge order matters. Until the DNS change lands, www.thenational.academy/mcp
is still claimed upstream — it currently returns a 200 HTML page titled "Oak
Curriculum MCP (HTTP)" from a non-OWA origin, and that exact URL is the OAuth
protected-resource identifier for the MCP server (RFC 8707 audience). Merging
this before the server has moved would leave the route shadowed at best.
Preview deployments are unaffected, as they serve from a different host.
