+++ src/app/(core)/mcp/Components/McpAssistants.tsx
@@ -11,10 +11,7 @@ import type { PortableTextComponents } from "@portabletext/react";
 import { McpSection } from "./McpSection";
 import { McpTryButton } from "./McpTryButton";
 
-import {
-  mcpAssistants,
-  type McpAssistant,
-} from "@/app/(core)/teachers/mcp/mcpContent";
+import { mcpAssistants, type McpAssistant } from "@/app/(core)/mcp/mcpContent";
 import { PortableTextWithDefaults } from "@/components/SharedComponents/PortableText";
 
 const McpAssistantCard = ({
