+++ src/app/(core)/mcp/Components/McpExternalLink.tsx
