+++ src/app/(core)/mcp/Components/McpSection.tsx
