+++ src/app/(core)/mcp/Components/McpTryButton.tsx
