+++ src/app/(core)/mcp/Components/McpVideoPlaceholder.tsx
