+++ src/app/(core)/mcp/Components/McpView.stories.tsx
