+++ src/app/(core)/mcp/Components/McpView.test.tsx
@@ -13,7 +13,7 @@ import {
   mcpOutputWarning,
   mcpResponsibleUse,
   mcpSupport,
-} from "@/app/(core)/teachers/mcp/mcpContent";
+} from "@/app/(core)/mcp/mcpContent";
 import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
 
 const render = renderWithProviders();
