+++ src/app/(core)/mcp/mcpContent.ts
