+++ src/app/(core)/mcp/page.tsx
